Lot of updates. Can now turn off drawing of lines to increase performance and updated python to create correct graphs with the data.

+++ toy_entropic_gravity/src/toy/entropic/gravity/Chord.java
@@ -2,10 +2,10 @@ package toy.entropic.gravity;
 
 public class Chord {
 	
-	private float X1, X2, Y1, Y2;
+	private double X1, X2, Y1, Y2;
 	private boolean excluded;
 	
-	public Chord(float x1, float y1, float x2, float y2) {
+	public Chord(double x1, double y1, double x2, double y2) {
 		X1 = x1;
 		X2 = x2;
 		Y1 = y1;
@@ -14,12 +14,12 @@ public class Chord {
 		excluded = false;
 	}
 	
-	public float getGradient() {
+	public double getGradient() {
 		return (Y2 - Y1) / (X2 - X1);
 	}
 	
 	// Y = m(X - i) + j where x = 0
-	public float getIntercept() {
+	public double getIntercept() {
 		return Y1 - getGradient() * X1;
 	}
 	
@@ -31,37 +31,37 @@ public class Chord {
 		excluded = e;
 	}
 	
-	public float getX1() {
+	public double getX1() {
 		return X1;
 	}
 	
-	public float getX2() {
+	public double getX2() {
 		return X2;
 	}
 	
-	public float getY1() {
+	public double getY1() {
 		return Y1;
 	}
 	
-	public float getY2() {
+	public double getY2() {
 		return Y2;
 	}
 	
 	// returns the x-coordinate of the midpoint of the chord
-	public float getX() {
+	public double getX() {
 		return (X2 - X1) / 2 + X1;
 	}
 	
 	// returns the y-coordinate of the midpoint of the chord
-	public float getY() {
+	public double getY() {
 		return (Y2 - Y1) / 2 + Y1;
 	}
 	
-	public float getDX() {
+	public double getDX() {
 		return X2 - X1;
 	}
 	
-	public float getDY() {
+	public double getDY() {
 		return Y2 - Y1;
 	}
